Removing older ANF prototypes
Leaving only SimpleANF, which is really just AST + smart constructors doing online normalization.
Other forms of ANF are way too complex and buggy, if we want them to have nice algebraic properties.
In the end, it's probably not worth it. Let's try to make SimpleANF as efficient as possible instead.