Sound ref checking on early return + hygiene fix in extraction
Ref checking was not performed when early return constructed code to be
returned. This is fixed.

Introduced `matchedSymbol` to hygienically match binders; still has
potentially surprising behaviors to be documented

Also, now ascribing types that have been narrowed after a rewriting.
This is mainly to avoid ending up in a situation where, e.g., a
transformed lambda is now typed Int=>Nothing but still used as if it was
an Int=>Int.