Optimized vanilla pattern-matching perf by caching the extractor
The caching is done if no term is inserted into the extractor.
If types are inserted (via explicit $$-insertion or implicit resolution),
these types are used as keys in the caching.
Nothing special is done about extracted terms and types.

Note that these do not affect top-level `rewrite`s, which already
store their extractors and thus do not rebuild them on every rewriting.
However, it does affect inner `rewrite` invocations!