[WIP] Pure Expression Rewriting
When the result of a multi-statement pattern is pure and could thus 
appear later (after some more statements) in the matched body.

This seems too hard to reason about, as it makes many things
surprisingly non-local. Sometimes an optim will apply on the wrong
scope  because of the added flexibility.
Also, it seems to sometimes even prevent some usual optimizations from
applying. (see Sieve/JoinLinesComplex)