Typing hole types as type members of local objects, instead of traits
This has the advantage that the type extracted for `$t` now appears 
as `t.Typ` instead of just `t` (a syntax not actually valid — that the 
user cannot use).
On the other hand, when an extracted type is extruded (taken out of its 
scope, for example returned from the pattern matching where it was 
extracted), it will no more appear as just `<extruded type>`, but as
something like ParentClass.t.Typ.