Make term-level `rewrite` use bottom-up order by default; closes #3
Bottom-up rewriting is safer (less prone to divergence) and often more
natural, so it’s a better default behavior.

Note that this is potentially a breaking change, though I expect that
most code that uses `rewrite` will still work.

Currently some tests are broken; they will be fixed in the next commit.