Implement support for first-class variable symbols
This is a big feature with far-reaching changes in core part of the
Squid system. We had to adapt the insertion and extraction mechanisms
to be able to insert and extract a new data type, and the quasi-
embedding mechanism to properly reflect the new semantics, in terms of
path-dependent context requirements and capture.
In addition, the IR implementations have to change their behavior with
respect to extracted binders.

This is a breaking change, because of name-based `subs` not working
anymore with extracted binders (which now have a new meaning).
Also had to adapt ASTReinterpreter to handle bindings/variables better.
The replacement for term.subs is varsymbol.substitute, which usually
requires explicit type arguments – a macro helper to regroup the two
features is needed to make the interface nicer to use.

Had to adapt the `rewrite` macro, which is closely dependent on the
shape of code generated by extractor quasiquotes.

Used rather ad-hoc fixes to some macro-specific problems with path-
dependent Variable.Ctx type, especially in `rewrite` expressions

Fixed and updated the `example` and `sc-backend` projects
The fix to the `sc-backend` IR implementation of binder extraction is
rather uncertain/tentative, and not well-tested