Support for embedding abstract types
This is to support using types in quasiquotes that are not classes,
such as opaque types for which only bounds are know or even no bounds
at all. Apparently we also somehow support higher-kinded types.

This is an improvement in particular because it means we infer less
type tags due to bailing on unknown types – which doesn’t really scale,
because you can’t embed Foo[T] if you only have a CodeType[T] but Foo is
abstract and would need a type tag (it came up with @embed in dbStage).

Importantly, before embedding an abstract type we first look for
implicit evidence, to support modular metaprogramming (see tests).