Fix recurrent and subtle type extraction problem by using type ranges
In the presence of declaration-site variance, singleton types and GADTs,
soundly extracting types turns out to be a rather arduous task.
The previous scheme was too approximative and bailed in surprising
cases, leading to things not matching that should obviously match – for
example, see test("Invariant Singleton Types Not Blocking Extraction").
We now use more flexible extracted _type ranges_ as extraction
artifacts for types. We still conserve a notion of variance (encoding in
what sort of context the type was matched), but use it to materialize
the range into a simple type (whereas it was previously used to do
extracted type merging).