New implementation of scheduling, suporting lazy values
Lazy values are created when a pure subexpression is duplicated in
several places, yet we do not know if it will ever be computed
because it only occurs in a lambda or a by-name argument passed
to a function with unknown semantics.

The squid.utils.Lazy data type is used to virtualize lazy values,
which has the advantage of being faster because not thread-safe.
In order for scheduling to be idempotent, I had to add a special
case in SimpleANF to remove squid.utils.Lazy redexes.

With this commit, I also renamed Occurrence to the more appropriate
ExecStats, and changed the mechanism of scheduling to avoid
performing repeated substitutions and recomputations of the exec
stats; instead keeping track of what subexpresions are scheduled,
like in the original implementation but with several flaws corrected.
This approach has the additional advantage that it avoids spurious
let-binding due to ANF losing track of the effect of an expression
when we substituted part of it with an opaque variable.

Fixes #50.