Macro-based implementation of static/compile-time execution
The compileTime{.} macro executes arbitrary code at compile time and
stores it in a compilation-run-local cache, so that when further code
is executed in the same compilation run, it can reuse the previously-
computed values.

The CompileTime data type is used for creating implicits whose
semantics is to provide various Squid functionalities with compile-
time values, such as the name of a file in which to log compile-time
information.

This is a step taken towards solving #19.