Simplify implicit evidence in .run and .compile; promote Any contexts
We used to represent empty contexts as {}, because we used to represent
context requirements as structural types; but this had several drawbacks.
Instead, we now promote the use of Any for empty contexts, and reflect
that in the .run and .compile methods.
Legacy code can still use {}, but needs to import LegacyPredef for it.