Misc. fixes; make unapplyConst much more powerful; some tests get SOF
Also, we really can’t completely partially evaluate the advanced tests
unless we use `extractTopLevelCallArg = true`, in which case we just
explore way too many paths.
This is because of extractor blindness: the IR doesn’t know we want to
match a constant, and so it puts things into a Hole that it should not;
and even with a very advanced Const xtor, we still see paths where the
constant to extract is ambiguous.
This can be solved by using a pattern the IR can see; for example adding
a squid.lib.const library function, and extending pat-mat accordingly.