Require annotation or explicit % for cross-stage references
This paves the way towards allowing cross-quotation references. It is
necessary because there is no reliable way to know, inside an inner
quasiquote, if something is a cross-stage or cross-quotation reference.
This is because at this point the outer code has not been type checked
yet, so we don’t know, for example, whether it will be lifted to code
via an implicit conversion.