Drop error-prone usage of hardRewireTo; instead just mutate val names
The tricky thing is to preserve original names for let bindings while
showing understandable names in debug output. This is difficult because
let bindings can be recursive, so to tie the knot we use rewireTo which
makes a recursive def’s top-level node to an empty Box(Id,_) redirection
node; this is problematic because the debug output ignores such nodes
since it shows an inlined view of Box control structures, and the ‘true’
defining node for the recursive def will have an automatically-generated
name unless we reset it ourselves.