Fix graph lifting of nested recursive defs: insert appropriate pop nodes
The problem was that when nested recursive definitions went through
lambdas and then reached a recursive occurrence, we directly accessed
the reference, which meant that we were traversing the lambdas again and
then on accessing a captured variable we would _only_ pop the lambdas
once, not accounting for being traversing the def for the second time!
So we need to put pop nodes around recursive occurrences that appear
under lambdas... at least unless the definition is top-level.
This is not a problem for top-level definitions because these cannot
capture external lambda-bound variables. So it seems we don’t need to do
that for them (but it might be more consistent to do it).