Make beta reduction not duplicate apps, to avoid unrolling recursions
We used to duplicate applications all over the place, and that caused
very strange structures with unrolled recursions, which the scheduler
simply could not handle (it created cycles in the arguments passed to
calls and calls themselves).
Now, we use a more 'respectful' implementation of beta reduction that
reconstructs more involved branch structures to avoid that.