Fix subtle scheduling bug: processing rec params may trigger changes
Previously, we were not careful about processing parameters for
recursive calls, which could end up adding parameters to the parent
recursive scope, which we would then miss.
Now, everything is in a loop and reprocessed if such a change occurs.
I don’t have tests for this because the tricky case only happened in
some temporary code, in the middle of a change.